Fix more warnings

+++ miniapps/common/mesh_extras.cpp
@@ -260,7 +260,7 @@ void KershawTransformation::Eval(Vector &V, ElementTransformation &T,
 
    X = x;
 
-   int layer = x*6.0;
+   int layer = static_cast<int>(x*6.0);
    real_t lambda = (x-layer/6.0)*6;
 
    // The x-range is split in 6 layers going from left-to-left, left-to-right,
